+++ src/pharmpy/modeling/expressions.py
@@ -1,0 +1,35 @@
+def get_observation_expression(model):
+    """Gets the full symbolic expression for the observation according to the model
+
+    This function currently only support models without ODE systems
+
+    Parameters
+    ----------
+    model : Model
+        Pharmpy model object
+
+    Returns
+    -------
+    Expression
+        Symbolic expression
+
+    Examples
+    --------
+    >>> from pharmpy.modeling import load_example_model, get_observation_expression
+    >>> import sympy
+    >>> model = load_example_model("pheno_linear")
+    >>> expr = get_observation_expression(model)
+    >>> sympy.pprint(expr)
+    D_EPSETA1_2⋅EPS(1)⋅(ETA(2) - OETA₂) + D_ETA1⋅(ETA(1) - OETA₁) + D_ETA2⋅(ETA(2)
+     - OETA₂) + EPS(1)⋅(D_EPS1 + D_EPSETA1_1⋅(ETA(1) - OETA₁)) + OPRED
+    """
+    stats = model.statements
+    for i, s in enumerate(stats):
+        if s.symbol == model.dependent_variable:
+            y = s.expression
+            break
+
+    for j in range(i, -1, -1):
+        y = y.subs({stats[j].symbol: stats[j].expression})
+
+    return y